BSSH Fastq Copy Manager: Refactor Stack
Removed remnants of icav2 copy batch utility stack.

Use the new icav2 copy batch component inside the bssh fastq copy manager
