+++ config/constants.ts
@@ -109,21 +109,20 @@ export const icav2AccessTokenSecretName: Record<AppStage, string> = {
   [AppStage.PROD]: 'ICAv2JWTKey-umccr-prod-service-prod', // pragma: allowlist secret
 };
 
-
 /*
 Resources generated by the BSSH Fastq Copy Manager
- */
+*/
 
-export const bsshFastqCopyManagerSSMRoot = '/orcabus/bssh_icav2_fastq_copy_manager';
-export const bsshFastqCopyManagerSSMName = path.join(
-  bsshFastqCopyManagerSSMRoot,
-  'state_machine_name'
-);
-export const bsshFastqCopyManagerSSMArn = path.join(
-  bsshFastqCopyManagerSSMRoot,
-  'state_machine_arn'
-);
-export const bsshFastqCopyManagerSfnName = 'bssh_fastq_copy_manager_sfn';
+export const bsshFastqCopyManagerWorkflowName = 'bsshFastqCopy';
+export const bsshFastqCopyManagerWorkflowTypeVersion = '1.0.0';
+export const bsshFastqCopyManagerServiceVersion = '2024.05.15';
+
+export const bsshFastqCopyManagerReadyEventSource = 'orcabus.workflowmanager';
+export const bsshFastqCopyManagerEventSource = 'orcabus.bsshfastqcopymanager';
+export const bsshFastqCopyManagerEventDetailType = 'WorkflowRunStateChange';
+
+/*
+Resources generated by the BCLConvert InterOp QC pipeline
 
 /*
 External resources required by the ctTSO v2 Stack
