+++ config/stacks/bsshIcav2FastqCopyManager.ts
@@ -1,10 +1,13 @@
 import {
   AppStage,
   icav2AccessTokenSecretName,
-  bsshFastqCopyManagerSfnName,
-  bsshFastqCopyManagerSSMName,
-  bsshFastqCopyManagerSSMArn,
   eventBusName,
+  bsshFastqCopyManagerWorkflowName,
+  bsshFastqCopyManagerWorkflowTypeVersion,
+  bsshFastqCopyManagerServiceVersion,
+  bsshFastqCopyManagerReadyEventSource,
+  bsshFastqCopyManagerEventSource,
+  bsshFastqCopyManagerEventDetailType,
 } from '../constants';
 import { BsshIcav2FastqCopyManagerConfig } from '../../lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/deploy/stack';
 
@@ -12,11 +15,13 @@ export const getBsshIcav2FastqCopyManagerStackProps = (
   stage: AppStage
 ): BsshIcav2FastqCopyManagerConfig => {
   return {
-    icav2CopyBatchUtilityStateMachineName: bsshFastqCopyManagerSfnName,
-    bsshIcav2FastqCopyManagerStateMachineName: bsshFastqCopyManagerSfnName,
-    bsshIcav2FastqCopyManagerStateMachineNameSsmParameterPath: bsshFastqCopyManagerSSMName,
-    bsshIcav2FastqCopyManagerStateMachineArnSsmParameterPath: bsshFastqCopyManagerSSMArn,
+    detailType: bsshFastqCopyManagerEventDetailType,
+    icav2TokenSecretId: icav2AccessTokenSecretName[stage],
+    internalEventSource: bsshFastqCopyManagerEventSource,
+    serviceVersion: bsshFastqCopyManagerServiceVersion,
+    triggerLaunchSource: bsshFastqCopyManagerReadyEventSource,
+    workflowName: bsshFastqCopyManagerWorkflowName,
+    workflowVersion: bsshFastqCopyManagerWorkflowTypeVersion,
     eventBusName: eventBusName,
-    icav2JwtSecretsManagerPath: icav2AccessTokenSecretName[stage],
   };
 };
